fix: send env file as secret
