+++ cli/common.py
@@ -148,3 +148,20 @@ def clean_modules(mod_dir):
                     generated_file = f"{mod_dir}/{path}/{sub_path}"
                     print(f"deleting {generated_file}")
                     os.remove(generated_file)
+
+
+def format_lambda_output(
+    json_response: str, json_output: bool, external_duration_sec: float, engine: str
+):
+    response = json.loads(json_response)
+    # enrich the event with the external invoke duration
+    response.setdefault("context", {})
+    response["context"]["external_duration_sec"] = external_duration_sec
+    response["context"]["engine"] = engine
+    if json_output:
+        return json.dumps(response)
+    else:
+        output = ""
+        for key in ["parsed_queries", "context", "resp", "logs"]:
+            output += f"{key.upper()}\n{response.get(key, '')}\n\n"
+        return output
