+++ cli/requirements.txt
@@ -4,3 +4,4 @@ invoke==1.7.1
 requests==2.28.1
 google-cloud-bigquery==3.3.5
 awslambdaric==2.0.4
+python-dotenv==0.21.0
