+++ docker/cli/Dockerfile
@@ -84,8 +84,9 @@ FROM cli-base as lambda
 COPY . /repo
 ENV REPO_DIR=/repo
 ENV CALLING_DIR=/tmp/
+ENV L12N_PLUGINS=lambdacli
 
-ENTRYPOINT [ "/repo/l12n", "lambdaric.run-bootstrap" ]
+ENTRYPOINT [ "/repo/l12n", "lambdacli.run-bootstrap" ]
 
 
 FROM cli-base as cli
